requirements.txt
asgiref==3.7.2
Django==3.2.25
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pytz==2024.2
sqlparse==0.4.4
typing_extensions==4.7.1
